kenfuse/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-Limiter==2.9.2  # Compatible with Python 3.8
python-dotenv==1.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
requests==2.31.0
reportlab==4.0.4
qrcode==7.4.2
redis==4.6.0  # Compatible version
